# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "TakeIn Resturant",
    "description": "A Food Delivering App"
  },
  "servers": [
    {
      "url": "http://takein-prod-public-alb-373215541.ap-southeast-1.elb.amazonaws.com/api/v1"
    },
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [{
    "bearerAuth": []
  }],
  "paths": {
...
    "/order": {
      "post": {
        "tags": ["ORDER"],
        "summary": "Order",
        "requestBody": {
          "description": "Order Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/ORDER"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "Your order successfully placed!",
            "schema": {
              "$ref": "#/definitions/ORDER"
            }
          },
          "400": {
            "description": "There is some error while placing order!."
          }
        }
      },
      "get": {
        "tags": ["ORDER"],
        "summary": "Fetching Previous Orders",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully fetched previous orders.",
            "schema": {
              "$ref": "#/definitions/ORDER"
            }
          },
          "400": {
            "description": "Error in fetching previous orders."
          }
        }
      }
    },
    "/cookie-maker/order/{id}": {
      "put": {
        "summary": "Update Order Status",
        "tags": ["COOKIE_MAKER"],
        "requestBody": {
          "description": "Order Status Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderStatus": {
                    "type": "string",
                    "example": "Accepted"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Order Object Id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated Order Status"
          },
          "400": {
            "description": "Failed to update order status"
          },
          "404": {
            "description": "Failed. Order Id not found"
          }
        }
      },
      "get": {
        "tags": ["COOKIE_MAKER"],
        "summary": "To get Order details.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Order Project id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Order Details"
          },
          "404": {
            "description": "Error in fetching Order Details."
          }
        }
      }
    },
    "/cookie-maker/order/list": {
      "get": {
        "tags": ["COOKIE_MAKER"],
        "summary": "Order List",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully fetched orders."
          },
          "400": {
            "description": "Error in fetching previous orders."
          }
        }
      }
    },
    "/cookie-maker/orders-summary": {
      "get": {
        "tags": ["COOKIE_MAKER"],
        "summary": "Orders Summary",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully fetched orders."
          },
          "400": {
            "description": "Error in fetching previous orders."
          }
        }
      }
    },
    "/cookie-maker/order/list?status={status}": {
      "get": {
        "tags": ["COOKIE_MAKER"],
        "summary": "Order List with Order Status",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "description": "Pass Order Status in String",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "You have successfully fetched orders."
          },
          "400": {
            "description": "Error in fetching previous orders."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ORDER": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "deliveryType": {
              "type": "string",
              "example": "Delivery"
            },
            "payWith": {
              "type": "string",
              "example": "Cash"
            }
          }
        }
      }
    },
...
    "id": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
...
  }
}